requirements.txt
Django==5.0.4
django-allauth==0.61.1
django-apscheduler==0.6.2
django-ckeditor==6.7.1
django-filter==24.2
celery==5.4.0
redis==5.0.4
